+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- integrity.Validator: Ability to get ModRevision from a validated result (#23).
+
 ### Changed
 
 ### Fixed